pyproject.toml
[project]
name = "modern-data-stack"
version = "0.1.0"
description = "Embedded Modern Data Stack — a serverless, local-first ELT pipeline boilerplate"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "dlt[parquet]>=1.0",
    "duckdb>=1.0",
    "polars>=1.0",
    "pyarrow>=17.0",
    "python-dotenv>=1.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[tool.setuptools.packages.find]
include = ["transform*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
